# server/src/main/resources/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Forum
  version: "0.1"
  description: Forum - API
tags:
  - name: Post
paths:
  /posts:
    get:
      tags:
        - Post
      operationId: getPosts
      security:
        - userAuth: []
      parameters:
        - $ref: "#/components/parameters/PageQueryParameter"
        - $ref: "#/components/parameters/SizeQueryParameter"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostModel"
        "401":
          $ref: "#/components/responses/Unauthorized"
servers:
  - url: /api/v1
components:
  parameters:
    PageQueryParameter:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeQueryParameter:
      name: size
      in: query
      schema:
        type: integer
        minimum: 10
        maximum: 30
        default: 10
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  securitySchemes:
    userAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    PostModel:
      type: object
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 600
        detail:
          type: string
        instance:
          type: string
          format: uri
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string